Fixed cmd_parse
cmd_parse now properly splits a line by spaces.
implemented str_n_arg() to return the number or words in a line.
CmdSequence member variable cmd_block(char**) is now always NULL
terminated. This means that a CmdSequence's cmd_block can easily be
passed as a parameter for a call to an execvX function