Added The Ability to run commands asynchronously
Adding a '&' as an arguement for a command will cause it to be ran
asynchronously when passed to cmd_exec_async.

feature is not complete yet. commands can run asynchronously but still
use the same stdin and stdout of the parent process.

the feature is not yet supported  by cmd_execute();